requirements.txt
discord.py
noise>=1.2.2
lxml>=4.5.1
pypng>=0.0.20